addfriend fix

+++ back-end/src/services/friendlist/friendlist.service.ts
@@ -69,6 +69,11 @@ export class FriendlistService {
       username: username,
       friendname: friendname
     };
+
+    if (!this.isFriend(username, friendname)) {
+      return;
+    }
+
     console.log("addfriend");
     await this.friendrequestService.removeRequest(username, friendname); // remove request
     const newEntry: FriendList = this.friendlistRepository.create(query);
